requirements.txt
Flask==2.3.3
python-nmap==0.7.1
scapy==2.5.0
matplotlib==3.7.2
networkx==3.1
pydot==1.4.2
python-dotenv==1.0.0